+++ previous-versions/commerce/src/codatcommerce/locations.py
@@ -3,7 +3,8 @@
 from .sdkconfiguration import SDKConfiguration
 from codatcommerce import utils
 from codatcommerce.models import errors, operations, shared
-from typing import Optional
+from jsonpath import JSONPath
+from typing import Any, Dict, Optional
 
 class Locations:
     r"""Retrieve standardized data from linked commerce platforms."""
@@ -102,8 +103,24 @@ class Locations:
             '5XX'
         ]))
         content_type = http_res.headers.get('Content-Type')
+        
+        def next_func() -> Optional[operations.ListLocationsResponse]:
+            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            next_cursor = JSONPath("").parse(body)
 
-        res = operations.ListLocationsResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
+            if len(next_cursor) == 0:
+                return None
+            next_cursor = next_cursor[0]
+
+            return self.list(
+                request=operations.ListLocationsRequest(
+                    company_id=request.company_id,
+                    connection_id=request.connection_id,
+                ),
+                retries=retries,
+            )
+
+        res = operations.ListLocationsResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res, next=next_func)
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):