+++ previous-versions/commerce/src/codatcommerce/models/operations/list_locations.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import locations as shared_locations
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -20,6 +20,7 @@ class ListLocationsRequest:
 
 @dataclasses.dataclass
 class ListLocationsResponse:
+    next: Callable[[], Optional[ListLocationsResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)