+++ previous-versions/commerce/src/codatcommerce/models/operations/list_orders.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import orders as shared_orders
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -28,6 +28,7 @@ class ListOrdersRequest:
 
 @dataclasses.dataclass
 class ListOrdersResponse:
+    next: Callable[[], Optional[ListOrdersResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)