+++ previous-versions/commerce/src/codatcommerce/models/operations/list_tax_components.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import taxcomponents as shared_taxcomponents
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -28,6 +28,7 @@ class ListTaxComponentsRequest:
 
 @dataclasses.dataclass
 class ListTaxComponentsResponse:
+    next: Callable[[], Optional[ListTaxComponentsResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)