+++ previous-versions/commerce/RELEASES.md
@@ -627,3 +627,13 @@ Based on:
 - [python v0.1.0] previous-versions/commerce
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-commerce/0.1.0 - previous-versions/commerce
+
+## 2023-09-12 00:17:42
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.1] previous-versions/commerce
+### Releases
+- [PyPI v0.1.1] https://pypi.org/project/codat-commerce/0.1.1 - previous-versions/commerce